crimson/*: rename errorator-loop.h into errorator-utils.h
(cherry picked from commit 682e4ff4322a895a0bdb6c9cc0d00a516499baad)
